+++ src/main/kotlin/csense/kotlin/not/highlighter/NotNameAnnotator.kt
@@ -1,34 +1,35 @@
 package csense.kotlin.not.highlighter
 
 import com.intellij.lang.annotation.*
-import com.intellij.openapi.editor.markup.*
 import com.intellij.openapi.util.*
 import com.intellij.psi.*
 import csense.idea.base.bll.kotlin.*
-import csense.kotlin.extensions.*
+import csense.idea.base.bll.string.*
 import csense.kotlin.not.highlighter.bll.*
-import csense.kotlin.not.highlighter.bll.StringBll.Companion.notText
 import csense.kotlin.not.highlighter.settings.*
-import org.jetbrains.kotlin.lexer.*
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.*
-import java.awt.*
 
 class NotNameAnnotator : Annotator {
 
-    private val settings = NotHighlighterSettings.instance
-
+    private val settings by lazy {
+        NotHighlighterSettings.instance
+    }
 
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
         if (settings.mayNotHighlight(element)) {
             return
         }
-
         when (element) {
             is KtNamedDeclaration -> element.highlight(holder)
             is KtNameReferenceExpression -> element.highlight(holder)
             is KtPrefixExpression -> element.highlight(holder)
+            is KtIsExpression -> element.highlight(holder)
+            is KtBinaryExpression -> element.highlight(holder)
         }
+    }
+
+    private fun KtBinaryExpression.highlight(holder: AnnotationHolder) {
+        highlightRange(range = operationReference.textRange, holder = holder)
     }
 
     private fun KtNamedDeclaration.highlight(holder: AnnotationHolder) {
@@ -40,58 +41,52 @@ class NotNameAnnotator : Annotator {
     }
 
     private fun KtPrefixExpression.highlight(holder: AnnotationHolder) {
-        if (operationToken != KtTokens.EXCL) {
-            return
-        }
-        val elementRange = TextRange(
-            /* startOffset = */ textRange.startOffset,
-            /* endOffset = */ textRange.startOffset + exclamationMark.length
-        )
-        highlightRange(range = elementRange, holder = holder)
+        highlightRange(range = textRangeOfOperator(), holder = holder)
+    }
+
+    private fun KtIsExpression.highlight(holder: AnnotationHolder) {
+        highlightRange(range = textRangeOfOperator(), holder = holder)
     }
 
 
     private fun highlightNotIn(
         element: PsiElement,
-        holder: AnnotationHolder,
-        length: Int = notText.length
+        holder: AnnotationHolder
     ) {
-        val notStart = element.indexOfNotOrNull() ?: return
-        val realNotStart = element.textRange.startOffset + notStart
-        highlightRange(
-            range = TextRange(
-                /* startOffset = */ realNotStart,
-                /* endOffset = */ realNotStart + length
-            ),
-            holder = holder
-        )
+        val startOffsetOfText = element.textRange.startOffset
+
+        element.text.camelCase.forEachCamelCaseWord { stringStartIndex, string ->
+            if (!string.isNotText()) {
+                return@forEachCamelCaseWord
+            }
+            val notTextRange = textRangeForNotText(startOffset = startOffsetOfText + stringStartIndex)
+            highlightRange(
+                range = notTextRange,
+                holder = holder
+            )
+        }
     }
+
+    private fun String.isNotText(): Boolean {
+        return equals(notText, ignoreCase = true)
+    }
+
+    private fun textRangeForNotText(startOffset: Int) = TextRange(
+        /* startOffset = */ startOffset,
+        /* endOffset = */ startOffset + notText.length
+    )
 
     private fun highlightRange(range: TextRange, holder: AnnotationHolder) {
-        val settings = NotHighlighterSettings.instance
-        holder.newSilentAnnotation(
-            /* severity = */ HighlightSeverity.INFORMATION
-        ).range(
-            /* range = */ range
-        ).enforcedTextAttributes(
-            TextAttributes(
-                /* foregroundColor = */  settings.foregroundColor,
-                /* backgroundColor = */ settings.backgroundColor,
-                /* effectColor = */ null,
-                /* effectType = */ null,
-                /* fontType = */ settings.fontSettings()
+        holder
+            .newSilentAnnotation(
+                /* severity = */ HighlightSeverity.INFORMATION
             )
-        ).create()
-    }
-
-
-    private fun PsiElement.indexOfNotOrNull(): Int? {
-        return text.bll.indexOfNotOrNull()
+            .range(range)
+            .enforcedTextAttributes(settings.toTextAttributes())
+            .create()
     }
 
     companion object {
-        const val exclamationMark = "!"
+        const val notText: String = "not"
     }
 }
-
-