+++ src/test/testData/NotCases.kt
@@ -1,6 +1,7 @@
+<info descr="null">@<info descr="null">file</info>:Suppress</info>("unused", "ControlFlowWithEmptyBody", "KotlinConstantConditions")
 package csense.kotlin.<info descr="null">not</info>.test
 
-class <info descr="null"><info descr="null">Not</info>Cases</info>(
+class <info descr="null">NotCases</info>(
     <info descr="null">private</info> val <info descr="null">thisValShouldBehighlighted<info descr="null">Not</info></info>: <info descr="null">Boolean</info>
 ) {
 
@@ -26,4 +27,33 @@ class <info descr="null"><info descr="null">Not</info>Cases</info>(
         return <info descr="null">a</info> + <info descr="null"><info descr="null">not</info>B</info>.<info descr="null">toString</info>()
     }
 
+    fun <info descr="null">highlightesIsCorrectly</info>(){
+        val <info descr="null">x</info>: <info descr="null">Any</info>? = "asd"
+        if(<info descr="null">x</info> is <info descr="null">String</info>){
+            //shouldnt be highligted
+        }
+        if(<info descr="null">x</info> <info descr="null">!is</info> <info descr="null">Int</info>){
+            //should be highligted
+        }
+    }
+
+
+    fun <info descr="null">multiple<info descr="null">Not</info>Are<info descr="null">Not</info></info>() {
+        val <info descr="null">x</info>: <info descr="null">Any</info>? = null
+        if (<info descr="null">x</info> <info descr="null">!is</info> <info descr="null">String</info> || <info descr="null">x</info> <info descr="null">!=</info> <info descr="null">Unit</info>) {
+
+        }
+        if (<info descr="null">x</info> <info descr="null">!=</info> <info descr="null">Unit</info>) {
+
+        }
+        if (<info descr="null">x</info> == <info descr="null">Unit</info>) {
+
+        }
+        if (<info descr="null">x</info> <info descr="null">!in</info> <info descr="null">setOf</info><<info descr="null">Any</info>>()) {
+
+        }
+        if (<info descr="null">x</info> in <info descr="null">setOf</info><<info descr="null">Any</info>>()) {
+
+        }
+    }
 }